# src/build/tsoa/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"URL": {
				"type": "string",
				"description": "The **`URL`** interface is used to parse, construct, normalize, and encode URL.\n\n[MDN Reference](https://developer.mozilla.org/docs/Web/API/URL)"
			},
			"ImageDTO": {
				"properties": {
					"id": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"url": {
						"$ref": "#/components/schemas/URL"
					},
					"height": {
						"type": "number",
						"format": "double"
					},
					"width": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id",
					"title",
					"url",
					"height",
					"width"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PaginatedResult_ImageDTO_": {
				"properties": {
					"data": {
						"items": {
							"$ref": "#/components/schemas/ImageDTO"
						},
						"type": "array"
					},
					"total": {
						"type": "number",
						"format": "double"
					},
					"offset": {
						"type": "number",
						"format": "double"
					},
					"limit": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"data",
					"total"
				],
				"type": "object",
				"additionalProperties": false
			},
			"HealthCheck": {
				"properties": {
					"uptime": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					},
					"timestamp": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"uptime",
					"message",
					"timestamp"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "image-uploader",
		"version": "1",
		"contact": {}
	},
	"paths": {
		"/images/{id}": {
			"get": {
				"operationId": "GetImage",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ImageDTO"
										}
									],
									"nullable": true
								}
							}
						}
					},
					"404": {
						"description": "Image not found"
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/images": {
			"get": {
				"operationId": "GetImages",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PaginatedResult_ImageDTO_"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "title",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"default": 0,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "SaveImages",
				"responses": {
					"200": {
						"description": "OK"
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "width",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "height",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"file": {
										"type": "string",
										"format": "binary"
									},
									"title": {
										"type": "string"
									}
								},
								"required": [
									"file",
									"title"
								]
							}
						}
					}
				}
			}
		},
		"/healthz": {
			"get": {
				"operationId": "Healthz",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HealthCheck"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}